+++ legwork/__init__.py
@@ -1,1 +1,2 @@
 from . import evol, psd, snr, source, strain, utils, visualisation
+from ._version import __version__